Make it possible to cross-compile this code to powerpc and s390x
This required removing the ring crate from the dependency tree, since it
compiles C (or asm?) code and doesn't work on those platforms. That meant
removing octocrab entirely and changing reqwest to use openssl instead of
rustls.

In order to keep the result binary static, we use the `vendored` feature for
the openssl crate, which compiles a new openssl locally and statically links
it.