Only run integration tests on Linux x86-64, macOS, and Windows
Getting these running on other archs would be challenging. First, we'd have to
use `cross build` when building the ubi that's run for tests. Second, the
tests would have to be adjusted to account to account for every platform with
a release for one of the projects that's used in these tests.