Fix asset matching for jq & mkcert
The jq releases use "linux32" and "linux64" as the OS/arch, which the matcher
did not look for.

For mkcert, the issue was that the release names include a version, like
"mkcert-v1.4.4-linux-arm64". This was matched as having an extesion of
".4-linux-arm64", and was then rejected for not having an extension ubi can
handle.

Now it will check to see if the asset name contains something like a version
and if so, it will check if the name has an "extension" that matches the last
bit after the dot in the version. If it does, this is not considered a real
extension and so it is accepted.