Switch from the `tar` crate to `binstall-tar`
This fixes an issue with untarring tarballs that use the GNU sparse
format. The tarball for the ubi Darwin x86-64 release uses the GNU sparse
format, which isn't supported by the tar crate
(https://github.com/alexcrichton/tar-rs/issues/295). Switching to the
binstall-tar fork of the tar crate fixes this.