Refactor asset picker to operate on `Keys<...>` where the keys are asset names
It never looked at the `Asset.url` field, so having it work with an owned `Vec<Asset>` was
pointless. This is step 1 of adding support for checksums, since to do that we have to preserve the
original list of assets, so we can find the checksums file after we pick the release asset we want
to download.