Make `ubi` look for an executable with a partial match of the project name if there's no exact match
This matches executables that start with the project name. On Windows, it still requires that they
end in `.bat` or `.exe`.

The implementation ended up being a bit gross because for tar files we have to re-open the file to
extract the match. See the code comments for more details on why.

This also adds docs on how `ubi` finds the executable inside an archive file.