Fix check for whether archive has a single top-level dir with `--extract-all`
Previously, this would extract the archive into the install root, _then_ check whether there was a
top-level dir. But if the install root already had files, this check always returned false. This
breaks the case where you want to install into an existing tree, like `~/.local`. For example, if
the archive contains a tree like `foo-v1.2/bin/foo`, `foo-v1.2/share/foo.toml`, etc., you want it to
end up as `~/.local/bin/foo`, etc.

This changes the code to extract to a temp dir, do the check for a single top-level dir, and then
_move_ everything to the install root, possibly omitting the single top-level dir if needed.